[refactor/auth] auth, user 폴더 구조 변경
