+++ src/main/java/com/hanghae/newsfeed/auth/dto/request/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package com.hanghae.newsfeed.user.dto.request;
+package com.hanghae.newsfeed.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
