+++ src/main/java/com/hanghae/newsfeed/auth/security/jwt/JwtTokenType.java
@@ -1,4 +1,4 @@
-package com.hanghae.newsfeed.security.jwt;
+package com.hanghae.newsfeed.auth.security.jwt;
 
 public enum JwtTokenType {
     ACCESS, REFRESH
