--- src/main/java/com/hanghae/newsfeed/user/controller/AuthController.java
@@ -1,36 +1,0 @@
-package com.hanghae.newsfeed.user.controller;
-
-import com.hanghae.newsfeed.user.dto.request.LoginRequestDto;
-import com.hanghae.newsfeed.user.dto.request.SignupRequestDto;
-import com.hanghae.newsfeed.user.dto.response.LoginResponseDto;
-import com.hanghae.newsfeed.user.dto.response.SignupResponseDto;
-import com.hanghae.newsfeed.user.service.UserService;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/auth")
-public class AuthController {
-    private final UserService userService;
-
-    // 회원 가입
-    @PostMapping("/signup")
-    public ResponseEntity<SignupResponseDto> signup(@RequestBody @Valid SignupRequestDto requestDto) {
-        return ResponseEntity.status(HttpStatus.OK).body(userService.signup(requestDto));
-    }
-
-    // 로그인
-    @PostMapping("/login")
-    public ResponseEntity<LoginResponseDto> login(@RequestBody LoginRequestDto requestDto) {
-        return ResponseEntity.status(HttpStatus.OK).body(userService.login(requestDto));
-    }
-}
